Add basic functionality of filtergin items with tags
Use & to check the if the item tag contains desired tag. Tags are 
represented in bit arrays, which is integer here. TagIdDic.json store 
the key-value pairs of tags and its bit representation. The 
FilterWithTags() function check through all the items meta data, which 
are stored in ItemTagDic.json, to see if the item contains. The time 
complexity is linear to the item count. Could have better way.